backend/requirements/base.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
docutils==0.18.1
drf-yasg==1.20.0
email-validator==1.1.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PyYAML==6.0
pyparsing==3.0.8
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
